Port old `test/language` tests into rspec...
In Ruby 1.9 the Test::Unit library was replaced by MiniTest.  This changes
some "esoteric" parts of the code, sadly including the parts of the library we
depend on to confine certain tests to only some systems.

Instead of investing heavily in rebuilding that facility over the new MiniTest
library it makes sense to just check for and either destroy or port the old
tests to the new rspec harness.

This is our desired future - that only one test framework exist - and we might
as well take full advantage of this opportunity to clean up our legacy code.

It also drops tests that duplicate testing already present in the spec suite.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>