Port Puppet[:path] test from Test::Unit to RSpec
This test checks that changing the setting results in changing the process
environment.  Not the most awesome test, but nothing else checks either.

It also highlights that we didn't preserve the process environment over tests,
which is bad, because that is a shared, persistent resource that will change
state in response to test actions.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>