Use PID-based ports to test web server listening.
When running more than one instance of the spec suite concurrently on a single
machine, the hard-coded listening ports in some of the integration tests could
lead to failures - when multiple suites hit the same tests at the same time.

This uses a port number based off the PID, which helps ensure we don't get
spurious failures from concurrency.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>