Port the provider tests from Test::Unit.
The provider tests exercise a whole pile of code that wasn't covered by the
specs.  This ports those tests over, and extends them to do further
combination testing in a few places.

These are very repetitive, table driven tests - but so is the underlying
feature we are trying to test.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>