Port the base service provider tests from Test::Unit.
The `base` service provider is terribly anæmic, but we had some tests, and it
does serve as the parent for almost every other provider.  So, port those
tests over.

Also, add a dash of cross-platform support by invoking the current Ruby
interpreter to do file system manipulation things - rather than leaning on
Unix tools that are not available everywhere.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>