Remove module-repository option from faces documention
The module_repository will still work since it's in Puppet's defaults,
it just won't be documented in the face help which is desired behavior
until we have better definition about how we will treat remote
repository interaction.

Also getting rid of the use of the ENV PUPPET_FORGE way to set
repository in upgrade.  It was completely inconsistent with the other
way to set repository and I assume was only added for debugging.