(#13326) Support `versionRequirement` in module metadata
Without this patch the module tool does not support using
`versionRequirement` to specify module dependency info. This results in
the latest version of the module dependency to be selected regardless
of the users intent.

This patch resolves this issue by falling back to `versionRequirement`
in the absence of `version_requirement` when setting the
version requirement of the dependency.

This patch includes updates tests.