Disable module tool caching behavior
The module installer was caching the tarballs for installed modules so
that if you tried to reinstall them, it used the already downloaded
tarball.  Unfortunately, it's very easy to delete a released version of
a module on the forge and reupload a modified module with the same
version number.  This means that anyone who downloaded the old content
will use their cached version, and be confused as to why they don't get
the new content, even after they uninstall or delete the installed
module.

This is a hacky way to disable the caching behavior, but long term the
code should be reworked to not do caching at all.  The tarball should be
put in a tmp directory on install, and cleaned up afterwards.  However,
being so close to release, we're going to do this short term fix.