Make new scoping look through inherited scopes
Previous to this commit, Puppet would look through a given scope
hierarchy and give deprecation warnings when a value was found in either
an inherited class scope or included class scope, when it should only
give the warning in the case of the included class scope.

This commit makes the new-scope behavior also examine inherited scopes,
though continuing to ignore included scopes.