Ruby 1.9.3-p0 support in spec tests.
Ruby 1.9.3-p0 will invoke `to_ary` on every element of an Array when
`Array#flatten` is invoked, which Ruby 1.8.7, and Ruby 1.9.3-p125 will not.

This leads to a failure in this test because our stub objects explode when you
invoke that method, rather than the normal Ruby behaviour.  Just stubbing that
method to return `nil` will do the right thing.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>