(#12392) Created Windows eventlog message resource dll
This commit adds the ability to build a message resource dll used to
display localized eventlog messages on Windows. Windows eventlog expects
that each log event has a unique id, which can then be localized in a
resource dll, one for each locale. However, puppet does not yet support
this, see #11076. So this commit defines three puppet event ids
corresponding to the three levels of Windows events that we support
(info, warn, and error).

In order to build the dll, you need the Windows SDK installed that
contains the mc, rc, and link utilities.