(#12392) Add eventlog destination and log to it by default on Windows
Previously, puppet logged by default to syslog or a file.

This commit changes puppet to log to syslog, the eventlog, or a file, in
that order. Puppet uses `Puppet.features.eventlog?` to determine if the
Windows eventlog is available on the current platform (which requires
the win32-eventlog gem). If the feature is false, then the eventlog
destination is not created, and the next suitable destination is used.