Fix Rails schema setup on Windows.
There were several problems with the StoreConfig Rails schema building code on
Windows, starting with the use of `/dev/null` to discard migration output.

Unfortunately, this was hard to debug - it broke tests silently because of the
hacking on $stdout, and didn't report anything about failures.

Now, instead of discarding the output we capture it in memory, and report any
exception we encounter along with the complete textual output at the time of
failure.

This improves the ability to debug errors in this code, along with making it
work transparently on any platform.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>