Forcibly remove temporary files in tests.
We have some test helpers, and they allow us to create scratch files and
directories that will be politely destroyed at the end of the run.

We also have some tests that result in permissions on files being changed; we
still want those destroyed, and turning on "force" mode for the delete
operation gives us what we need with no real cost - we have plenty of other
safety factors in place.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>