Recover better when `maxgroups` is not implemented.
Puppet tries to bump up the maximum number of supplementary groups for the
process, which works only on Unix-like platform, in C-like implementations of
Ruby.

Previously there was a blacklist of platforms where this didn't work, which
means that as new platforms come along we failed in new and exciting ways.

Instead we can just rescue any failure from invoking the method and carry on,
which works for any failure mode including "not implemented" nicely.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>