(#13642) Move search notice message
Currently the PMT search action assumes the searcher implementation will
connect to the HTTP server defined by
`Puppet.settings[:module_repository]` when searching for modules on the
Forge. The server defined by `Puppet.settings[:module_repository]` is then
used in a message printed on the console: `Searching #{server} …`. This
is a little too close to the implementation and the notice has been
moved closer to where the search is actually taking place.