fix bug in Util#symbolizehash!
The method Puppet::Util.symbolizehash! was modifying a hash
during iteration; this is (sanely, thankfully) not legal
in ruby 1.9.  This has been broken in our code for quite
some time, but we hadn't noticed because nothing was calling
this method.  A recent commit introduced code that calls it,
highlighting the bug, which is fixed by this commit.