Store Rails logs on physical disk when testing.
On Unix, we stored the Rails log from our StoreConfigs scratch database in
`/dev/null`, which is reasonable enough - we don't generally care for them,
and that works.

On Windows this is harder because their NUL device is magic, and it is hard to
configure the Rails logs to actually go there.  The system loves to interpret
it as a random file relative to the current directory and do the wrong thing
in various test scenarios.

Ultimately, it doesn't matter if we write the logs to physical disk, and we
already had a scratch directory for the physical database - so we can throw
the logs there too.

Just another reason to use a tmpfs to back /tmp, I guess.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>