Removing use of deprecated execution methods
Using the mixins for bringing in the execution methods wasn't
working to bring in the non-deprecated Puppet::Util::Execution methods.
Because of the way that these methods are declared trying to understand
what methods were coming in from where was very hard to trace and so
I decided to go with explicit delegation.