Manually load Win32{API,OLE} for Ruby 1.9.
In Ruby 1.8.7 on Windows the Win32API and Win32OLE objects were automatically
loaded; in Ruby 1.9 there is a general move to get rid of them, and presumably
one step along the path is that they no longer automatically load.

Explicitly loading them along with other gems we depend on helps solve that
problem for us automatically.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>