(maint) Remove unnecessary step from puppet module build help
Without this patch the `puppet help module build` command says to make
sure the module directory is in the form of <username>-<modulename>.
This is an unnecessary step because it isn't actually necessary anymore.
It's a problem because we want to make things appear to be as simple as
they actually are, not more complex than they need to be.

I just tested this with a module existing at /vagrant/modules/jenkins
(no symlinks) using the command: `puppet module build .`

Everything worked out fine.  Puppet correctly reads the Modulefile and
figures out the author name.

This is a documentation only change.