(#13204) Don't ignore missing PATH.augnew files
The original fix for #13204 may have masked other potential bugs if the
PATH.augnew file was missing.  It simply tested for the file existance and not
only when duplicate save events occurred.

This change de-duplicates the list of save events instead, so if a bug appeared
where PATH.augnew was genuinely missing, the error wouldn't be squashed.