(Maint) Fix order-dependent test failure
Previously, the test assumed that the agent already had a signed
certificate and would fail trying to connect to the master to retrieve a
catalog. But this didn't work if the agent didn't have a cert.

This commit changes the test to use a conf directory, whose parents
don't exist. As a result, puppet will immediately fail to start, but
report the error, including the path we specified. This way it's not
dependent on the SSL state, and it doesn't affect the SSL-ness of other
tests.