Update puppet.spec for rpm packaging
As el4 is now EOL, Ruby 1.8.1 doesn't need to be the minimum Ruby. 1.8.5 is
available in el5, so that is a sane new minimum. This commit also removes a
macro that was needed for building rpms on platforms earlier than el5. It also
updates the changelog and version for the latest release of puppet, 2.7.13.