(#13645) Add test cases to tagmail_spec
If puppet is configured to use tagmail as a report, it should not send
an email if no message match any tag in tagmail.conf

The former behaviour was that puppet opened an smtp connection, did not
send any mail and then closed the connection again leading to errors on
the smtp server. The added tests now prove that the send method is never
even called when no message match any positive tag.