(#3757) Remove 'anonymous' functionality from lockfiles
The agent lockfile mechanisms had a concept of "anonymous"
built-in to them.  This was necessary to help attempt to
disambiguate the meaning of the lockfile when we were
using it for two purposes (administrative-disabling vs.
run-in-progress).  Since we now have two separate lockfiles,
this is no longer necessary... and it makes the code
and API confusing.  This commit removes it, and the
resulting API feels much cleaner and more straightforward.

The "AnonymousFilelock" class needs to be renamed, but
I haven't done that yet.  My next commit is going to
transform it into a JSON-based lockfile class, so I'll
rename it as part of that commit.