Use conditional pending to block out "fails_on_windows" tests.
A whole bunch of tests scattered through the system fail on Windows, around
features that are not supported on that platform.  (They are things that only
the master does, which an agent-only platform doesn't need to support.)

These were tagged `fails_on_windows` to allow filtering them from rspec runs,
which is great, but doesn't actually communicate nearly as much useful
information as it would if we used the "conditionally pending" facilities that
rspec has supported since 2.3.

That gives us two key things: one, it works automatically based on our
knowledge of the platform, which means you can't forget to turn off failing
tests.

Two, it means that if the test starts unexpectedly passing we also get a
failure, since we should respond to "works when it shouldn't" as seriously as
"fails when it shouldn't".

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>